Moves SubmanDaemon
